+++ tests/PhpStan/phpstan-bootstrap.php
@@ -18,6 +18,11 @@ class_alias(
 );
 
 class_alias(
+    \OxidEsales\Eshop\Application\Component\UserComponent::class,
+    \OxidEsales\SecurityModule\Captcha\Shop\UserComponent_parent::class
+);
+
+class_alias(
     \OxidEsales\Eshop\Application\Model\User::class,
     \OxidEsales\SecurityModule\Shared\Model\User_parent::class
 );